Update chenho's code to right directory
