--- src/chenho/Answer.java
@@ -1,6 +1,0 @@
-package src;
-
-public class Answer{
-    public String message;
-    public String tag;
-}
