--- src/chenho/Item.java
@@ -1,16 +1,0 @@
-package src;
-
-/**
- * 아이템이 소유한 플레이어에게 가산해줄 능력치를 정의한 클래스
- * @author Chanho Park
- */
-
-public class Item
-    {
-        public int p_power=0;
-        public int p_dex=0;
-        public int p_int=0;
-        public int p_health=0;
-        public int p_mental=0;
-        public String name;
-    }
