--- src/chenho/Player.java
@@ -1,14 +1,0 @@
-package src;
-
-public class Player{
-    int power;
-    int dexterity;
-    int intelligence;
-    int health;
-    int mental;
-    TileType pos;
-    int money;
-    int status;
-    Item[] items;
-    int energy;
-}
