--- src/chenho/Tile.java
@@ -1,15 +1,0 @@
-package src;
-
-/**
- * 타일 객체 모델, 각 타일은 자신의 소환된 몬스터나 차원문, 내부 이벤트를 관리한다.
- * @author Se-Ok Jeon
- * @version 1.0
- */
-
-public class Tile {
-
-    private boolean summoned_portals;
-    private Monster summoned_monsters;
-    private String name;
-
-}
