+++ src/main/java/com/Answer.java
@@ -1,0 +1,6 @@
+package src;
+
+public class Answer{
+    public String message;
+    public String tag;
+}
