+++ src/main/java/com/Monster.java
@@ -42,4 +42,7 @@ public class Monster
     private int monster_result;
     private MonsterType name;
 
+    public decreaseHealth(int damage_from_player){
+        health -= damage_from_player;
+    }
 }
